Migrate search and more buttons to the breadcrumb toolbar.
