+++ client/localization/elements/sc-universal-action-items/en.json
@@ -1,0 +1,6 @@
+{
+  "en": {
+      "searchTooltip": "Search SuttaCentral",
+      "Search": "Search"
+  }
+}
