+++ client/localization/elements/sc-universal-action-items/pl.json
@@ -1,0 +1,6 @@
+{
+  "pl": {
+    "searchTooltip": "Wyszukaj w SuttaCentral",
+    "Search": "Szukaj"
+  }
+}
