+++ client/localization/elements/sc-universal-action-items/pt.json
@@ -1,0 +1,6 @@
+{
+  "pt": {
+      "searchTooltip": "Buscar no SuttaCentral",
+      "Search": "Busca"
+  }
+}
