+++ client/localization/elements/sc-universal-action-items/zh.json
@@ -1,0 +1,6 @@
+{
+  "zh": {
+      "searchTooltip": "搜索SuttaCentral",
+      "Search": "搜索"
+  }
+}
