+++ client/redux-store.js
@@ -69,7 +69,6 @@ const initialState = {
     displayToolButton: false,
     displayInfoButton: false,
     displayViewModeButton: true,
-    displaySCSiteTitle: true,
     staticPagesToolbarDisplayState: {
       displayFirstToolbar: true,
       displaySecondToolbar: false,
@@ -163,17 +162,15 @@ const reducer = (state, action) => {
         case 'CHANGE_DISPLAY_TOOL_BUTTON_STATE':
             return Object.assign({}, state, { displayToolButton: action.displayToolButton });
         case 'CHANGE_DISPLAY_INFO_BUTTON_STATE':
-          return Object.assign({}, state, { displayInfoButton: action.displayInfoButton });
+            return Object.assign({}, state, { displayInfoButton: action.displayInfoButton });
         case 'CHANGE_DISPLAY_VIEW_MODE_BUTTON_STATE':
-          return Object.assign({}, state, { displayViewModeButton: action.displayViewModeButton });
-        case 'CHANGE_DISPLAY_SC_SITE_TITLE_STATE':
-          return Object.assign({}, state, { displaySCSiteTitle: action.displaySCSiteTitle });
+            return Object.assign({}, state, { displayViewModeButton: action.displayViewModeButton });
         case 'CHANGE_NAV_STATE':
             return Object.assign({}, state, { navState: action.navState });
         case 'CHANGE_CURRENT_NAV_POSITION_STATE':
-          return Object.assign({}, state, { currentNavPosition: action.currentNavPosition });
+            return Object.assign({}, state, { currentNavPosition: action.currentNavPosition });
         case 'CHANGE_STATIC_PAGES_TOOLBAR_DISPLAY_STATE':
-          return Object.assign({}, state, { staticPagesToolbarDisplayState: action.staticPagesToolbarDisplayState });
+            return Object.assign({}, state, { staticPagesToolbarDisplayState: action.staticPagesToolbarDisplayState });
         default:
             return state;
     }
